+++ Thunderstore/CHANGELOG.md
@@ -1,3 +1,7 @@
+### v1.1.0
+- rewritten as a preloader
+- more reactive source registration
+
 ### v1.0.1
 - tweaks and slight performance improvement